Move silkenweb_base::window and document to silkenweb
