+++ packages/silkenweb/src/node/element.rs
@@ -19,7 +19,6 @@ use futures_signals::{
     signal_vec::{always, SignalVec, SignalVecExt},
     CancelableFutureHandle,
 };
-use silkenweb_base::document;
 use silkenweb_signals_ext::value::{Executor, RefSignalOrValue, SignalOrValue, Value};
 use wasm_bindgen::{JsCast, JsValue};
 
@@ -27,7 +26,7 @@ use self::child_vec::ChildVec;
 use super::{ChildNode, Node, ResourceVec};
 use crate::{
     attribute::Attribute,
-    clone,
+    clone, document,
     dom::{
         private::{DomElement, DomText, EventStore, InstantiableDomElement},
         DefaultDom, Dom, Hydro, InDom, InstantiableDom, Template, Wet,
